Fixed updating hobbies

+++ totally-legit-horoscopes-api/Models/LifeNumber.cs
@@ -5,7 +5,7 @@ namespace totally_legit_horoscopes_api.Models
     public class LifeNumber
     {
         [Key]
-        public int LifeNumberInt { get; set; }
+        public long LifeNumberInt { get; set; }
         public string Description { get; set; }
     }
 }
